chore: move tablerow to separate component
